move fs tests to their own file
